Rename player to entity
