+++ src/core/entity.hpp
